Utils: move Enum to package enum
