+++ discord/guild_const.go
@@ -1,7 +1,7 @@
 package discord
 
 import (
-	"github.com/diamondburned/arikawa/utils/json/nullable"
+	"github.com/diamondburned/arikawa/utils/json/enum"
 )
 
 // Guild.MaxPresences is 5000 when it's 0.
@@ -54,12 +54,12 @@ const (
 )
 
 // ExplicitFilter is the explicit content filter level of a guild.
-type ExplicitFilter nullable.Enum
+type ExplicitFilter enum.Enum
 
 var (
 	// NullExplicitFilter serialized to JSON null.
 	// This should only be used on nullable fields.
-	NullExplicitFilter ExplicitFilter = nullable.EnumNull
+	NullExplicitFilter ExplicitFilter = enum.Null
 	// NoContentFilter disables content filtering for the guild.
 	NoContentFilter ExplicitFilter = 0
 	// MembersWithoutRoles filters only members without roles.
@@ -69,23 +69,23 @@ var (
 )
 
 func (f *ExplicitFilter) UnmarshalJSON(b []byte) error {
-	i, err := nullable.EnumFromJSON(b)
+	i, err := enum.FromJSON(b)
 	*f = ExplicitFilter(i)
 
 	return err
 }
 
 func (f ExplicitFilter) MarshalJSON() ([]byte, error) {
-	return nullable.EnumToJSON(nullable.Enum(f)), nil
+	return enum.ToJSON(enum.Enum(f)), nil
 }
 
 // Notification is the default message notification level of a guild.
-type Notification nullable.Enum
+type Notification enum.Enum
 
 var (
 	// NullNotification serialized to JSON null.
 	// This should only be used on nullable fields.
-	NullNotification Notification = nullable.EnumNull
+	NullNotification Notification = enum.Null
 	// AllMessages sends notifications for all messages.
 	AllMessages Notification = 0
 	// OnlyMentions sends notifications only on mention.
@@ -93,21 +93,21 @@ var (
 )
 
 func (n *Notification) UnmarshalJSON(b []byte) error {
-	i, err := nullable.EnumFromJSON(b)
+	i, err := enum.FromJSON(b)
 	*n = Notification(i)
 
 	return err
 }
 
-func (n Notification) MarshalJSON() ([]byte, error) { return nullable.EnumToJSON(nullable.Enum(n)), nil }
+func (n Notification) MarshalJSON() ([]byte, error) { return enum.ToJSON(enum.Enum(n)), nil }
 
 // Verification is the verification level required for a guild.
-type Verification nullable.Enum
+type Verification enum.Enum
 
 var (
 	// NullVerification serialized to JSON null.
 	// This should only be used on nullable fields.
-	NullVerification Verification = nullable.EnumNull
+	NullVerification Verification = enum.Null
 	// NoVerification required no verification.
 	NoVerification Verification = 0
 	// LowVerification requires a verified email
@@ -124,13 +124,13 @@ var (
 )
 
 func (v *Verification) UnmarshalJSON(b []byte) error {
-	i, err := nullable.EnumFromJSON(b)
+	i, err := enum.FromJSON(b)
 	*v = Verification(i)
 
 	return err
 }
 
-func (v Verification) MarshalJSON() ([]byte, error) { return nullable.EnumToJSON(nullable.Enum(v)), nil }
+func (v Verification) MarshalJSON() ([]byte, error) { return enum.ToJSON(enum.Enum(v)), nil }
 
 // Service is used for guild integrations and user connections.
 type Service string
