+++ utils/json/enum/enum.go
@@ -1,20 +1,20 @@
-package nullable
+package enum
 
 import "strconv"
 
-// EnumNull is the value used to represent JSON null.
+// Null is the value used to represent JSON null.
 // It should never be used as a value, as it won't get serialized as such.
-const EnumNull = -1
+const Null = -1
 
 // Enum is a nullable version of a uint8.
 // Enum values should only consist of positive values, as negative values are reserved for internal constants, such as
-// EnumNull.
+// Null.
 // This also mean that only 7 of the 8 Bit will be available for storage.
 type Enum int8
 
 // Int8ToJSON converts the passed Enum to a byte slice with it's JSON representation.
-func EnumToJSON(i Enum) []byte {
-	if i == EnumNull {
+func ToJSON(i Enum) []byte {
+	if i == Null {
 		return []byte("null")
 	} else {
 		return []byte(strconv.Itoa(int(i)))
@@ -22,11 +22,11 @@ func EnumToJSON(i Enum) []byte {
 }
 
 // Int8FromJSON decodes the Enum stored as JSON src the passed byte slice.
-func EnumFromJSON(b []byte) (Enum, error) {
+func FromJSON(b []byte) (Enum, error) {
 	s := string(b)
 
 	if s == "null" {
-		return EnumNull, nil
+		return Null, nil
 	} else {
 		i, err := strconv.ParseUint(s, 10, 7)
 		return Enum(i), err
